Clean up interactions views
