+++ src/interactions/urls.py
@@ -1,47 +1,59 @@
 from django.urls import path
 
-from . import views
+from interactions.views import (
+    bookmarks,
+    comment_reactions,
+    comments,
+    page_reactions,
+    tag_subscriptions,
+)
 
 
 app_name = "interactions"
 
 urlpatterns = [
-    path("bookmark", views.bookmark, name="bookmark"),
     path(
-        "comment-reactions/<int:pk>/", views.react_to_comment, name="comment-reactions"
+        "comment-reactions/<int:pk>/",
+        comment_reactions.react_to_comment,
+        name="comment-reactions",
     ),
-    path("bookmark/<int:pk>/remove", views.remove_bookmark, name="bookmark-remove"),
-    path("bookmarks", views.bookmark_index, name="bookmark-index"),
-    path("page/<int:pk>/react", views.react_to_page, name="page-reactions"),
-    path("page/<int:pk>/comments", views.get_page_comments, name="get-page-comments"),
-    path("page/<int:pk>/comment", views.comment_on_page, name="comment-on-page"),
+    path("bookmark", bookmarks.bookmark, name="bookmark"),
+    path("bookmark/<int:pk>/remove", bookmarks.remove_bookmark, name="bookmark-remove"),
+    path("bookmarks", bookmarks.bookmark_index, name="bookmark-index"),
+    path("page/<int:pk>/react", page_reactions.react_to_page, name="page-reactions"),
     path(
-        "comment/<int:comment_id>/reply", views.reply_to_comment, name="reply-comment"
+        "page/<int:pk>/comments", comments.get_page_comments, name="get-page-comments"
     ),
-    path("comment/<int:pk>/hide", views.hide_comment, name="hide-comment"),
+    path("page/<int:pk>/comment", comments.comment_on_page, name="comment-on-page"),
+    path(
+        "comment/<int:comment_id>/reply",
+        comments.reply_to_comment,
+        name="reply-comment",
+    ),
+    path("comment/<int:pk>/hide", comments.hide_comment, name="hide-comment"),
     path(
         "comment/<int:comment_id>/edit",
-        views.edit_comment,
+        comments.edit_comment,
         name="edit-comment",
     ),
     path(
         "comment/<int:comment_id>/edit/form",
-        views.edit_comment_form,
+        comments.edit_comment_form,
         name="edit-comment-form",
     ),
     path(
         "comment/<int:comment_id>/",
-        views.get_comment,
+        comments.get_comment,
         name="get-comment",
     ),
     path(
         "tag/<int:tag_pk>/subscribe/",
-        views.subscribe,
+        tag_subscriptions.subscribe,
         name="subscribe-to-tag",
     ),
     path(
         "tag/<int:tag_pk>/unsubscribe/",
-        views.unsubscribe,
+        tag_subscriptions.unsubscribe,
         name="unsubscribe-from-tag",
     ),
 ]
