+++ src/interactions/views/bookmarks.py
@@ -1,0 +1,55 @@
+from django.http import (
+    HttpRequest,
+    HttpResponse,
+)
+from django.shortcuts import get_object_or_404
+from django.template.response import TemplateResponse
+from django.urls import reverse
+from django.views.decorators.http import require_http_methods
+from wagtail.models import Page
+
+from interactions.services import bookmarks as bookmarks_service
+
+
+@require_http_methods(["POST"])
+def bookmark(request: HttpRequest):
+    user = request.user
+
+    if request.method == "POST":
+        page_id = int(request.POST["page_id"])
+        page = get_object_or_404(Page, id=page_id)
+
+        bookmarks_service.toggle_bookmark(user, page)
+
+        is_bookmarked = bookmarks_service.is_page_bookmarked(user, page)
+
+        context = {
+            "post_url": reverse("interactions:bookmark"),
+            "user": user,
+            "page": page,
+            "is_bookmarked": is_bookmarked,
+        }
+
+        return TemplateResponse(
+            request,
+            "interactions/bookmark_page_input.html",
+            context,
+        )
+
+
+@require_http_methods(["DELETE"])
+def remove_bookmark(request: HttpRequest, *, pk):
+    bookmarks_service.remove_bookmark(pk, request.user)
+    return HttpResponse()
+
+
+@require_http_methods(["GET"])
+def bookmark_index(request: HttpRequest):
+    return TemplateResponse(
+        request,
+        "interactions/bookmark_index.html",
+        context={
+            "bookmarks": bookmarks_service.get_bookmarks(request.user),
+            "extra_breadcrumbs": [(None, "Your bookmarks")],
+        },
+    )
