+++ src/interactions/views/comments.py
@@ -3,7 +3,6 @@ from django.http import (
     HttpRequest,
     HttpResponse,
     HttpResponseForbidden,
-    JsonResponse,
 )
 from django.shortcuts import get_object_or_404, redirect
 from django.template.response import TemplateResponse
@@ -12,80 +11,10 @@ from django.views.decorators.http import require_http_methods
 from wagtail.models import Page
 
 from core import flags
-from core.models.tags import Tag
 from core.utils import flag_is_active
-from interactions.models import ReactionType
-from interactions.services import bookmarks as bookmarks_service
-from interactions.services import comment_reactions as comment_reactions_service
 from interactions.services import comments as comments_service
-from interactions.services import page_reactions as page_reactions_service
-from interactions.services import tag_subscriptions as tag_subscriptions_service
 from news.forms import CommentForm
 from news.models import Comment
-
-
-@require_http_methods(["POST"])
-def bookmark(request: HttpRequest):
-    user = request.user
-
-    if request.method == "POST":
-        page_id = int(request.POST["page_id"])
-        page = get_object_or_404(Page, id=page_id)
-
-        bookmarks_service.toggle_bookmark(user, page)
-
-        is_bookmarked = bookmarks_service.is_page_bookmarked(user, page)
-
-        context = {
-            "post_url": reverse("interactions:bookmark"),
-            "user": user,
-            "page": page,
-            "is_bookmarked": is_bookmarked,
-        }
-
-        return TemplateResponse(
-            request,
-            "interactions/bookmark_page_input.html",
-            context,
-        )
-
-
-@require_http_methods(["DELETE"])
-def remove_bookmark(request: HttpRequest, *, pk):
-    bookmarks_service.remove_bookmark(pk, request.user)
-    return HttpResponse()
-
-
-@require_http_methods(["GET"])
-def bookmark_index(request: HttpRequest):
-    return TemplateResponse(
-        request,
-        "interactions/bookmark_index.html",
-        context={
-            "bookmarks": bookmarks_service.get_bookmarks(request.user),
-            "extra_breadcrumbs": [(None, "Your bookmarks")],
-        },
-    )
-
-
-@require_http_methods(["POST"])
-def react_to_page(request: HttpRequest, *, pk):
-    page = get_object_or_404(Page, id=pk)
-    user = request.user
-
-    if request.method == "POST":
-        reaction_type = ReactionType(request.POST["reaction_type"])
-        is_selected = request.POST.get("is_selected") == "true"
-
-        reacted_type = None if is_selected else reaction_type
-        page_reactions_service.react_to_page(user, page, reacted_type)
-
-    return JsonResponse(
-        {
-            "user_reaction": page_reactions_service.get_user_page_reaction(user, page),
-            "reactions": page_reactions_service.get_page_reaction_counts(page),
-        }
-    )
 
 
 @require_http_methods(["POST"])
@@ -167,27 +96,6 @@ def reply_to_comment(request: HttpRequest, *, comment_id):
     return comments_service.get_comment_response(request, comment)
 
 
-def react_to_comment(request: HttpRequest, *, pk):
-    comment = get_object_or_404(Comment, id=pk)
-    user = request.user
-
-    if request.method == "POST":
-        reaction_type = ReactionType(request.POST["reaction_type"])
-        is_selected = request.POST.get("is_selected") == "true"
-
-        reacted_type = None if is_selected else reaction_type
-        comment_reactions_service.react_to_comment(user, comment, reacted_type)
-
-    return JsonResponse(
-        {
-            "user_reaction": comment_reactions_service.get_user_comment_reaction(
-                user, comment
-            ),
-            "reactions": comment_reactions_service.get_comment_reaction_counts(comment),
-        }
-    )
-
-
 @require_http_methods(["POST"])
 def comment_on_page(request: HttpRequest, *, pk):
     page = get_object_or_404(Page, id=pk).specific
@@ -215,19 +123,3 @@ def hide_comment(request: HttpRequest, pk: int) -> HttpResponse:
     comments_service.hide_comment(comment)
 
     return comments_service.get_page_comments_response(request, comment.page)
-
-
-@require_http_methods(["POST"])
-def subscribe(request, *, tag_pk: int) -> HttpResponse:
-    user = request.user
-    tag = Tag.objects.get(pk=tag_pk)
-    tag_subscriptions_service.subscribe(tag=tag, user=user)
-    return redirect(reverse("tag_index", kwargs={"slug": tag.slug}))
-
-
-@require_http_methods(["POST"])
-def unsubscribe(request, *, tag_pk: int) -> HttpResponse:
-    user = request.user
-    tag = Tag.objects.get(pk=tag_pk)
-    tag_subscriptions_service.unsubscribe(tag=tag, user=user)
-    return redirect(reverse("tag_index", kwargs={"slug": tag.slug}))
